requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.0
django-apscheduler==0.6.2
django-filter==23.5
django-modeltranslation==0.18.11
djangorestframework==3.15.1
idna==3.6
kombu==5.3.5
Markdown==3.6
oauthlib==3.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.13.3
